+++ src/common/common.less
@@ -6,7 +6,7 @@
 html,
 body,
 #root {
-  min-height: 100%;
+  min-height: 100vh;
   overscroll-behavior-y: none;
   // height: 100%;
 }
